requirements.txt
typing-inspect==0.8.0
typing_extensions==4.5.0
openai
langchain
jupyter
python-dotenv
pydantic==1.10.8
